Badge becomes Indicator, Indicator becomes badge
