+++ src/components/Badge/index.test.js
@@ -1,17 +1,65 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { render } from '@testing-library/react'
+import '@testing-library/jest-dom/extend-expect'
 
 import Badge from '.';
 
+const SELECTOR_NOTIF = '[aria-label="Notification badge"]';
+const SELECTOR_COUNT = '[aria-label="Count badge"]';
+
 describe('Badge component', () => {
 
-  it('matches snapshot (default props)', () => {
-    const component = shallow(<Badge label="badge" />);
-    expect(component).toMatchSnapshot();
+  it('renders correctly as notif with label', () => {
+    const { container } = render(<Badge label="666" />);
+    const notif = container.querySelector(SELECTOR_NOTIF);
+    const count = container.querySelector(SELECTOR_COUNT);
+    expect(count).toBeFalsy();
+    expect(notif).toBeTruthy();
+    expect(notif).toHaveClass('color--white');
+    expect(notif).toHaveClass('bgColor--red');
+    expect(notif).not.toHaveClass('color--aqua');
+    expect(notif).not.toHaveClass('bgColor--aqua');
   });
 
-  it('matches snapshot (set all props)', () => {
-    const component = shallow(<Badge label="badge" size="xs" data-test="test" />)
-    expect(component).toMatchSnapshot();
+  it('renders correctly as notif without label', () => {
+    const { container } = render(<Badge />);
+    const notif = container.querySelector(SELECTOR_NOTIF);
+    const count = container.querySelector(SELECTOR_COUNT);
+    expect(count).toBeFalsy();
+    expect(notif).toBeTruthy();
+    expect(notif).toHaveClass('color--white');
+    expect(notif).toHaveClass('bgColor--red');
+    expect(notif).toHaveClass('shape--circle');
+    expect(notif).toHaveClass('fds-badge--dot');
+    expect(notif).toHaveClass('color--white');
+    expect(notif).not.toHaveClass('color--aqua');
+    expect(notif).not.toHaveClass('bgColor--aqua');
   });
+
+  it('renders correctly as count with label', () => {
+    const { container } = render(<Badge label="666" type="count" />);
+    const notif = container.querySelector(SELECTOR_NOTIF);
+    const count = container.querySelector(SELECTOR_COUNT);
+    expect(notif).toBeFalsy();
+    expect(count).toBeTruthy();
+    expect(count).toHaveClass('color--aqua');
+    expect(count).not.toHaveClass('color--white');
+    expect(count).not.toHaveClass('bgColor--red');
+    expect(count).not.toHaveClass('shape--circle');
+  });
+
+  it('renders correctly as count without label', () => {
+    const { container } = render(<Badge type="count" />);
+    const notif = container.querySelector(SELECTOR_NOTIF);
+    const count = container.querySelector(SELECTOR_COUNT);
+    expect(notif).toBeFalsy();
+    expect(count).toBeTruthy();
+    expect(count).toHaveClass('color--aqua');
+    expect(count).toHaveClass('bgColor--aqua');
+    expect(count).toHaveClass('shape--circle');
+    expect(count).toHaveClass('fds-badge--dot');
+    expect(count).toHaveClass('color--white');
+    expect(count).not.toHaveClass('bgColor--red');
+  });
+
 });
